move interfaces/tidebit_defi_background/json.ts to interfaces/ijson.ts
