--- src/interfaces/tidebit_defi_background/json.ts
@@ -1,3 +1,0 @@
-export type IJSON = string | number | boolean | null | IJSON[] | {[key: string]: IJSON};
-
-export type ILocale = {locale: string};
