+++ src/interfaces/tidebit_defi_background/locale.ts
@@ -1,0 +1,1 @@
+export type ILocale = {locale: string};
