+++ src/pages/404.tsx
@@ -5,7 +5,7 @@ import NavBar from '../components/nav_bar/nav_bar';
 import {serverSideTranslations} from 'next-i18next/serverSideTranslations';
 import Footer from '../components/footer/footer';
 import {AppContext} from '../contexts/app_context';
-import {ILocale} from '../interfaces/tidebit_defi_background/json';
+import {ILocale} from '../interfaces/tidebit_defi_background/locale';
 
 // TODO: layoutAssertion
 // TODO: AppInit
